character-set function added to return characters based on the name of the charSet
