+++ app/character-set-builder.js
@@ -1,20 +1,4 @@
-const {
-  ALPHABET_UPPERCASE,
-  ALPHABET_LOWERCASE,
-  DIGIT,
-  BINARY,
-  OCTAL,
-  HEX,
-  HEX_LOWER,
-  CHARSET_ALPHA,
-  CHARSET_ALPHA_LOWER,
-  CHARSET_DIGIT,
-  CHARSET_ALNUM,
-  CHARSET_BINARY,
-  CHARSET_OCTAL,
-  CHARSET_HEX,
-  CHARSET_HEX_LOWER
-} = require('./constants.js');
+const characterSet = require('./character-set.js');
 
 /**
  * This will generate a string of unique characters based on the options provided.
@@ -22,49 +6,10 @@ const {
  * @returns {string} The set of characters based on the options provided.
  */
 function characterSetBuilder(characterSetOptions) {
-
-  /**
-   * This will return the characters based on the character set name.
-   * @param {string} charSet This is the name of the character set.
-   * @returns {string} String of characters.
-   */
-  function characters(charSet) {
-    switch (charSet) {
-      case CHARSET_ALPHA:
-        return ALPHABET_UPPERCASE;
-      case CHARSET_ALPHA_LOWER:
-        return ALPHABET_LOWERCASE;
-      case CHARSET_DIGIT:
-        return DIGIT;
-      case CHARSET_ALNUM:
-        return `${ALPHABET_UPPERCASE}${ALPHABET_LOWERCASE}${DIGIT}`;
-      case CHARSET_BINARY:
-        return `${BINARY}`;
-      case CHARSET_OCTAL:
-        return `${OCTAL}`;
-      case CHARSET_HEX:
-        return `${HEX}`;
-      case CHARSET_HEX_LOWER:
-        return `${HEX_LOWER}`;
-      default:
-        const validCharSets = [
-          CHARSET_ALPHA,
-          CHARSET_ALPHA_LOWER,
-          CHARSET_DIGIT,
-          CHARSET_ALNUM,
-          CHARSET_BINARY,
-          CHARSET_OCTAL,
-          CHARSET_HEX,
-          CHARSET_HEX_LOWER
-        ];
-        throw new Error(`Invalid builtIn characterSet specified. Allowed values: ${validCharSets.join(', ')}`);
-    }
-  }
-
   const {builtIn = [], custom = []} = characterSetOptions;
 
   const builtInCharacters = builtIn.reduce((chars, charSet) => {
-    return `${chars}${characters(charSet)}`;
+    return `${chars}${characterSet(charSet)}`;
   }, '');
 
   const customCharacters = custom.reduce((chars, charSet) => {
